update buffer create

+++ oz/src/oz/gfx/vulkan/vk_enums.h
@@ -6,6 +6,6 @@ namespace oz::gfx::vk {
 
 enum class ShaderStage : uint8_t { Vertex = 0x00000001, Fragment = 0x00000010, Compute = 0x00000020 };
 
-enum class BufferType : uint32_t { Vertex, Uniform, Index, Staging };
+enum class BufferType : uint8_t { Vertex, Uniform, Index, Staging };
 
 } // namespace oz::gfx::vk
